+++ src/arpes/fits/__init__.py
@@ -24,7 +24,6 @@ from .fit_models.misc import (
     LogRenormalizationModel,
 )
 from .fit_models.two_dimensional import EffectiveMassModel, Gaussian2DModel
-from .fit_models.x_model_mixin import XModelMixin, gaussian_convolve
 from .utilities import broadcast_model, result_to_hints
 
 if TYPE_CHECKING: